Duplicate time functions in app with different names
These functions are not desired in the libs API, but are used in both
the lib and the App.  They can't have the same name since this causes
duplicate symbols in statically link executable.
