+++ Source/App/EncApp/EbTime.h
@@ -1,0 +1,28 @@
+/*
+* Copyright(c) 2019 Intel Corporation
+* SPDX - License - Identifier: BSD - 2 - Clause - Patent
+*/
+
+#ifndef EbTime_h
+#define EbTime_h
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif // __cplusplus
+
+#define NANOSECS_PER_SEC ((uint32_t)(1000000000L))
+
+void StartTime(uint64_t *Startseconds, uint64_t *Startuseconds);
+void FinishTime(uint64_t *Finishseconds, uint64_t *Finishuseconds);
+void ComputeOverallElapsedTime(uint64_t Startseconds, uint64_t Startuseconds, uint64_t Finishseconds, uint64_t Finishuseconds, double *duration);
+void ComputeOverallElapsedTimeMs(uint64_t Startseconds, uint64_t Startuseconds, uint64_t Finishseconds, uint64_t Finishuseconds, double *duration);
+void Injector(uint64_t processedFrameCount, uint32_t injector_frame_rate);
+
+#ifdef __cplusplus
+}
+#endif // __cplusplus
+
+#endif // EbTime_h
+/* File EOF */
